prepare: use --no-gpg for bodhi updates download
See discussion in
https://pagure.io/fedora-ci/general/issue/536 . Earlier this year
I sent a patch to make bodhi updates download try and download
signed packages by default:
https://github.com/fedora-infra/bodhi/pull/5859 . However this is
actually a bit less reliable because it's possible for Koji not
to be able to find signed packages and bail out without getting
anything. There's going to be a "fall back to unsigned if signed
doesn't work" behavior but it's not all wired up yet.

As rmdepcheck doesn't need the packages to be signed, let's just
use --no-gpg to skip looking for signed packages and avoid any
problems.

Signed-off-by: Adam Williamson <awilliam@redhat.com>